Backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-dotenv==1.0.0
pydantic==2.5.2
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
pytest==7.4.4
httpx==0.26.0
supabase==2.0.3
websockets==12.0
flask-socketio==5.3.6
python-engineio==4.9.0
python-socketio==5.10.0
reportlab==4.0.7
pdfkit==1.0.0
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.3
joblib==1.3.2
transformers==4.36.2
torch==2.1.2
redis==5.0.1
celery==5.3.7
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-restx==1.2.0
flask-apispec==0.13.0
pytest-cov==4.1.0
black==23.12.1
flake8==7.0.0
mypy==1.8.0
